rename multi-select to select
