+++ src/domain/discounts/new/index.js
@@ -9,7 +9,7 @@ import Medusa from "../../../services/api"
 import AvailabilityDuration from "../../../components/availability-duration"
 import Button from "../../../components/button"
 import Pill from "../../../components/pill"
-import MultiSelect from "../../../components/molecules/multi-select"
+import MultiSelect from "../../../components/molecules/select"
 import Input from "../../../components/molecules/input"
 import Typography from "../../../components/typography"
 import ProductSelection from "../product-selection"
